improve API validation

+++ React/my-contactcs-project/api/src/app/controllers/ContactController.js
@@ -1,5 +1,5 @@
 const ContactsRepository = require("../repositories/ContactsRepository");
-
+const isValidUUID = require("../utils/isValidUUID");
 class ContactController {
   async index(req, res) {
     // Listar todos os registros
@@ -12,10 +12,15 @@ class ContactController {
   async show(req, res) {
     // Obter um registro
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: "Invalid contact id" });
+    }
+
     const contact = await ContactsRepository.findById(id);
 
+    // 404: Not Found
     if (!contact) {
-      // 404: Not Found
       return res.status(404).json({ error: "Contact not found" });
     }
 
@@ -30,17 +35,22 @@ class ContactController {
       return res.status(400).json({ error: "Name is required" });
     }
 
-    const contactExists = await ContactsRepository.findByEmail(email);
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: "Invalid category" });
+    }
 
-    if (contactExists) {
-      return res.status(400).json({ error: "This e-mail is already in use" });
+    if (email) {
+      const contactExists = await ContactsRepository.findByEmail(email);
+      if (contactExists) {
+        return res.status(400).json({ error: "This e-mail is already in use" });
+      }
     }
 
     const contact = await ContactsRepository.create({
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     res.status(201).json(contact);
@@ -51,28 +61,35 @@ class ContactController {
     const { id } = req.params;
     const { name, email, phone, category_id } = req.body;
 
-    const contactExists = await ContactsRepository.findById(id);
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: "Invalid contact id" });
+    }
 
-    if (!contactExists) {
-      // 404: Not Found
-      return res.status(404).json({ error: "User not found" });
+    if (category_id && !isValidUUID(category_id)) {
+      return res.status(400).json({ error: "Invalid category" });
     }
 
     if (!name) {
       return res.status(400).json({ error: "Name is required" });
     }
 
-    const contactByEmail = await ContactsRepository.findByEmail(email);
+    const contactExist = await ContactsRepository.findById(id);
+    if (!contactExist) {
+      res.status(404).json({ error: "Contact not found" });
+    }
 
-    if (contactByEmail && contactByEmail.id !== id) {
-      return res.status(400).json({ error: "This e-mail is already in use" });
+    if (email) {
+      const contactByEmail = await ContactsRepository.findByEmail(email);
+      if (contactByEmail && contactByEmail.id !== id) {
+        return res.status(400).json({ error: "This e-mail is already in use" });
+      }
     }
 
     const contact = await ContactsRepository.update(id, {
       name,
-      email,
+      email: email || null,
       phone,
-      category_id,
+      category_id: category_id || null,
     });
 
     res.json(contact);
@@ -81,6 +98,10 @@ class ContactController {
   async delete(req, res) {
     // Deletar um registro
     const { id } = req.params;
+
+    if (!isValidUUID(id)) {
+      return res.status(400).json({ error: "Invalid contact id" });
+    }
 
     await ContactsRepository.delete(id);
